pyproject.toml
[project]
name = "wikipedia"
version = "0.1.0"
description = "Wikipedia pageviews analytics system with multi-cloud deployment support"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # Core analysis dependencies
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "plotly>=5.14.0",
    "scipy>=1.10.0",
    "pywavelets>=1.4.1",
    "statsmodels>=0.14.0",
    "jupyter>=1.0.0",
    "ipykernel>=6.25.0",
    "boto3>=1.28.0",
    # Optional: Advanced analytics (for --use-enriched flag)
    "networkx>=3.0",
    "sparqlwrapper>=2.0.0",
    "requests-cache>=1.0.0",
    "python-louvain>=0.16",
    "rich>=13.0.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
evals = [
    "pyyaml>=6.0",
]
